Made a crude battle state machine
The battle state machine is found in the BoxCombat scene. The scene
includes an arena with a blue and red box in each corner, a button that
reads “Attack”, and text that has player and enemy health.
When the “Attack” button is pressed, the blue box does damage to the
red box, then the red box does damage to the blue box. The state
machine then waits for the button to be pressed again. When the red
box’s health reaches zero a message pops up that says “You win”.