Updated Combat
Put the moves into the player class with getters (and setters). Did the
same with EnemyPlayer. Enemy player now has a method where it gets a
random attack to attack.

Next I need to add the rest (health++) and a log for what attacks were
used. Also need to add an “energy” count so that the most expensive
move cannot just be done over and over.