Implemented an energy variable to the BasePlayer
If the BasePlayer does not have enough energy for an attack, those
attacks will disappear from the scene. To gain energy, the player must
rest. The player loses a turn, gains 2 energy points and takes damage
from the enemy.  Next step is to implement the same for the enemy
player.